Use docker-compose to manage the containers.
Updated postgres to the official 9.4 image
Use docker compose to manage the containers
Cleanup dockerfiles
Remove host volume for postgres data
Cleanup createdb script
Add recreatedb script
Don't run createdb on container start

+++ musicbrainz-dockerfile/scripts/createdb.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
-apt-get install -y wget
-rm -rf /media/dbdump/*
-wget -nd -nH -P /media/dbdump ftp://ftp.musicbrainz.org/pub/musicbrainz/data/fullexport/LATEST
-LATEST=$(cat /media/dbdump/LATEST)
-wget -r --no-parent -nd -nH -P /media/dbdump --reject "index.html*, mbdump-edit*, mbdump-documentation*" "ftp://ftp.musicbrainz.org/pub/musicbrainz/data/fullexport/$LATEST"
-pushd /media/dbdump && md5sum -c MD5SUMS && popd
-/musicbrainz-server/admin/InitDb.pl --createdb --import /media/dbdump/mbdump*.tar.bz2 --echo
+
+FETCH_DUMPS=$1
+
+if [[ $FETCH_DUMPS == "-fetch" ]]; then
+  echo "fetching data dumps"
+
+  apt-get install -y wget
+  rm -rf /media/dbdump/*
+  wget -nd -nH -P /media/dbdump ftp://ftp.musicbrainz.org/pub/musicbrainz/data/fullexport/LATEST
+  LATEST=$(cat /media/dbdump/LATEST)
+  wget -r --no-parent -nd -nH -P /media/dbdump --reject "index.html*, mbdump-edit*, mbdump-documentation*" "ftp://ftp.musicbrainz.org/pub/musicbrainz/data/fullexport/$LATEST"
+  pushd /media/dbdump && md5sum -c MD5SUMS && popd
+  /musicbrainz-server/admin/InitDb.pl --createdb --import /media/dbdump/mbdump*.tar.bz2 --echo
+elif [[ -a /media/dbdump/mbdump.tar.bz2 ]]; then
+  echo "found existing dumps"
+
+  /musicbrainz-server/admin/InitDb.pl --createdb --import /media/dbdump/mbdump*.tar.bz2 --echo
+else
+  echo "no dumps found or dumps are incomplete"
+  /musicbrainz-server/admin/InitDb.pl --createdb --echo
+fi
